pyproject.toml
[project]
name = "ucp-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "a2a-sdk>=0.3.22",
    "fastmcp>=2.14.0",
    "google-adk>=1.22.1",
]
